requirements.txt
attrdict==2.0.0
ecdsa==0.13
mock==1.0.1
pyOpenSSL==16.0.0
requests==2.7.0
six==1.9.0
wsgiref==0.1.2